requirements.txt
streamlit
python-docx
yake
